Рассчет статистики вынесен в отдельный хук
